+++ src/main/java/ua/foxminded/pskn/universitycms/model/user/Student.java
@@ -18,6 +18,6 @@ public class Student {
     @Id
     private Long userId;
 
-    private int groupId;
+    private Long groupId;
 
 }
